工作台数据 知识库 合同附件

+++ merchant-system/src/main/java/com/merchant/system/domain/ContractFee.java
@@ -19,8 +19,8 @@ public class  ContractFee extends BaseEntity
 {
     private static final long serialVersionUID = 1L;
 
-    /** 合同费用表id */
-    private Long id;
+    /** 费用表id */
+    private Integer id;
 
     private Integer receiverId;
